+++ src/com/lukxtech/parts/transmissionfamily/transmissions/dct/enums/DCTModelType.java
@@ -8,7 +8,7 @@ public enum DCTModelType {
     public static DCTModelType fromString(String type){
         return switch (type.toLowerCase().trim()){
             case "dct8" -> DCT8;
-            default -> throw new InvalidDCTModelTypeException("INVALID CVT MODEL TYPE! GOT: " + type + "\nPERMISSIVE: cvt6");
+            default -> throw new InvalidDCTModelTypeException("INVALID DCT MODEL TYPE! GOT: " + type + "\nPERMISSIVE: dct8");
         };
     }
 }