+++ CHANGELOG.md
@@ -1,7 +1,17 @@
 # Change Log
 
-> Version Information...
+> Version Information...<br>
 > 1.0.0.a - {Major Version}.{Minor Version}.{Bugs Fixed/Stuff Implemented}.{Version Type}
+---
+
+## 1.1.2.a
+## Added
+- Added DCT Transmission Implementation
+  - Check `src.com.lukxtech.parts.transmissionfamily.transmissions.dct.DualClutchTransmission.java`
+- Fixed Line issue with Changelog
+
+## Notes:
+Very small push. Mostly intended to close the Transmission Family Saga.
 
 ---